update plan selector to use mui dialog
